Move all print_* to modules/
