+++ src/modules/print_brightness.cc
@@ -14,11 +14,11 @@
 
 #include <exception>
 
-#include "utility.h"
-#include "process_configuration.h"
-#include "handle_click_events.h"
-#include "formatting/printer.hpp"
-#include "formatting/Conditional.hpp"
+#include "../utility.h"
+#include "../process_configuration.h"
+#include "../handle_click_events.h"
+#include "../formatting/printer.hpp"
+#include "../formatting/Conditional.hpp"
 #include "print_brightness.h"
 
 using swaystatus::Conditional;
