+++ src/modules/print_memory_usage.cc
@@ -13,12 +13,12 @@
 #include <exception>
 #include <string_view>
 
-#include "utility.h"
-#include "process_configuration.h"
-#include "handle_click_events.h"
-#include "formatting/printer.hpp"
-#include "formatting/LazyEval.hpp"
-#include "mem_size_t.hpp"
+#include "../utility.h"
+#include "../process_configuration.h"
+#include "../handle_click_events.h"
+#include "../formatting/printer.hpp"
+#include "../formatting/LazyEval.hpp"
+#include "../mem_size_t.hpp"
 #include "print_memory_usage.h"
 
 using namespace std::literals;
