+++ src/modules/print_network_interfaces.cc
@@ -6,11 +6,11 @@
 
 #include <exception>
 
-#include "process_configuration.h"
-#include "handle_click_events.h"
-#include "formatting/printer.hpp"
-#include "formatting/Conditional.hpp"
-#include "networking.hpp"
+#include "../process_configuration.h"
+#include "../handle_click_events.h"
+#include "../formatting/printer.hpp"
+#include "../formatting/Conditional.hpp"
+#include "../networking.hpp"
 #include "print_network_interfaces.h"
 
 using swaystatus::Conditional;
