+++ src/modules/print_sensors.cc
@@ -1,10 +1,10 @@
 #include <err.h>
 #include <exception>
 
-#include "sensors.hpp"
-#include "process_configuration.h"
-#include "handle_click_events.h"
-#include "formatting/printer.hpp"
+#include "../sensors.hpp"
+#include "../process_configuration.h"
+#include "../handle_click_events.h"
+#include "../formatting/printer.hpp"
 #include "print_sensors.h"
 
 using swaystatus::Sensors;
