+++ src/modules/print_time.c
@@ -6,9 +6,9 @@
 #include <time.h>
 #include <unistd.h>
 
-#include "process_configuration.h"
-#include "handle_click_events.h"
-#include "formatting/printer.hpp"
+#include "../process_configuration.h"
+#include "../handle_click_events.h"
+#include "../formatting/printer.hpp"
 #include "print_time.h"
 
 static const char * const module_name = "time";
